requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
crispy-bootstrap4==2022.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.7
django-crispy-forms==2.0
django-import-export==3.2.0
et-xmlfile==1.1.0
import-export==0.3.1
MarkupPy==1.14
mssql-django==1.2
mysql==0.0.3
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.1.2
pyodbc==4.0.35
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.3
tablib==3.4.0
tzdata==2023.2
xlrd==2.0.1
xlwt==1.3.0

